requirements.txt
appdirs==1.4.4
attrs==24.1.0
beautifulsoup4==4.13.3
CacheControl==0.14.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
coverage==7.8.0
cryptography==44.0.2
cssselect==1.2.0
dbus-python==1.4.0
filelock==3.18.0
GDAL==3.10.2
idna==3.10
lensfun==0.3.4
libtorrent==2.0.11
lockfile==0.12.2
lxml==5.3.2
mercurial==7.0.1
meson==1.8.0
msgpack==1.0.5
numpy==2.2.5
packaging==25.0
pillow==11.2.1
psutil==7.0.0
pycairo==1.28.0
pycparser==2.22
PyGObject==3.52.3
pyserial==3.5
pyxdg==0.28
requests==2.32.3
scour==0.38.2
six==1.17.0
soupsieve==2.6
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
urwid==2.6.16
wcwidth==0.2.13
wheel==0.45.1
zstandard==0.23.0
